+++ NAMSC_lib/src/Novel/Data/Visual/Scenery/Scenery.cpp
@@ -77,23 +77,23 @@ bool Scenery::operator==(const Scenery& obj) const noexcept
 void Scenery::serializableLoad(QDataStream& dataStream)
 {
 	dataStream >> backgroundAssetImageName_ >> musicPlaylist;
-	size_t displayedCharactersSize, displayedSceneryObjectsSize, soundsSize;
+	uint displayedCharactersSize, displayedSceneryObjectsSize, soundsSize;
 	dataStream >> displayedCharactersSize >> displayedSceneryObjectsSize >> soundsSize;
-	for (size_t i = 0; i != displayedCharactersSize; ++i)
+	for (uint i = 0; i != displayedCharactersSize; ++i)
 	{
 		Character character;
 		dataStream >> character;
 		addDisplayedCharacter(std::move(character));
 	}
 
-	for (size_t i = 0; i != displayedSceneryObjectsSize; ++i)
+	for (uint i = 0; i != displayedSceneryObjectsSize; ++i)
 	{
 		SceneryObject sceneryObject;
 		dataStream >> sceneryObject;
 		addDisplayedSceneryObject(std::move(sceneryObject));
 	}
 
-	for (size_t i = 0; i != soundsSize; ++i)
+	for (uint i = 0; i != soundsSize; ++i)
 	{
 		Sound sound;
 		dataStream >> sound;
@@ -106,7 +106,7 @@ void Scenery::serializableLoad(QDataStream& dataStream)
 void Scenery::serializableSave(QDataStream& dataStream) const
 {
 	dataStream << backgroundAssetImageName_ << musicPlaylist;
-	dataStream << displayedCharacters_.size() << displayedSceneryObjects_.size() << sounds_.size();
+	dataStream << static_cast<uint>(displayedCharacters_.size()) << static_cast<uint>(displayedSceneryObjects_.size()) << static_cast<uint>(sounds_.size());
 	for (const Character& character : displayedCharacters_)
 		dataStream << character;
 	for (const SceneryObject& sceneryObject : displayedSceneryObjects_)
